move conversion objects to conversion module
