+++ src/compas_rhino/conversions/_exceptions.py
