+++ src/compas_rhino/conversions/polyline.py
@@ -4,8 +4,8 @@ from __future__ import division
 
 import Rhino
 
-from compas_rhino.conversions import polyline_to_compas
-from compas_rhino.conversions import polyline_to_rhino
+from ._primitives import polyline_to_compas
+from ._primitives import polyline_to_rhino
 
 from ._geometry import RhinoGeometry
